+++ src/Demo/DemoModule.php
@@ -5,22 +5,19 @@ declare(strict_types=1);
 namespace App\Demo;
 
 use Aphiria\Application\Builders\IApplicationBuilder;
-use Aphiria\Application\IModule;
-use Aphiria\Framework\Application\AphiriaComponents;
+use Aphiria\Framework\Application\AphiriaModule;
 use Aphiria\Net\Http\HttpStatusCodes;
 use App\Demo\Binders\UserServiceBinder;
 
 /**
  * Defines the demo module
  */
-final class DemoModule implements IModule
+final class DemoModule extends AphiriaModule
 {
-    use AphiriaComponents;
-
     /**
      * @inheritdoc
      */
-    public function build(IApplicationBuilder $appBuilder): void
+    public function configure(IApplicationBuilder $appBuilder): void
     {
         $this->withBinders($appBuilder, new UserServiceBinder())
             ->withProblemDetails(